Did a bit more refactoring before new dynamic example (wip) started working
