+++ src/brom_drake/productions/__init__.py
@@ -2,9 +2,12 @@ from .ids import ProductionID
 from brom_drake.productions.debug.show_me import (
     ShowMeThisModel,
 )
-from brom_drake.productions.motion_planning.offline import (
+from brom_drake.productions.motion_planning.offline.kinematic import (
     ShelfPlanning1,
     ChemLab1,
+)
+from brom_drake.productions.motion_planning.offline.dynamic import (
+    ChemLab2,
 )
 
 __all__ = [
@@ -12,4 +15,5 @@ __all__ = [
     "ShowMeThisModel",
     "ShelfPlanning1",
     "ChemLab1",
+    "ChemLab2",
 ]
