+++ src/brom_drake/productions/types/motion_planning/__init__.py
@@ -1,3 +1,3 @@
-from .kinematic import KinematicMotionPlanningProduction
+from .offline.kinematic import KinematicMotionPlanningProduction
 
 __all__ = ["KinematicMotionPlanningProduction"]
