+++ src/brom_drake/utils/__init__.py
@@ -2,13 +2,15 @@ from .search import (
     find_all_systems_with_input_port,
     find_all_systems_with_output_port,
 )
-from .bool_to_vec_system import BoolToVectorSystem
+from .leaf_systems.bool_to_vec_system import BoolToVectorSystem
+from .leaf_systems.end_effector_wrench_calculator import EndEffectorWrenchCalculator
+from .leaf_systems.rigid_transform_to_vector_system import RigidTransformToVectorSystem
 from .constants import Performer, MotionPlan
 from .ground import AddGround, GroundShape
-from .rigid_transform_to_vector_system import RigidTransformToVectorSystem
 
 __all__ = [
     "BoolToVectorSystem",
+    "EndEffectorWrenchCalculator",
     "find_all_systems_with_input_port",
     "find_all_systems_with_output_port",
     "Performer",
