+++ src/brom_drake/utils/leaf_systems/__init__.py
@@ -1,5 +1,9 @@
+from .bool_to_vec_system import BoolToVectorSystem
 from .end_effector_wrench_calculator import EndEffectorWrenchCalculator
+from .rigid_transform_to_vector_system import RigidTransformToVectorSystem
 
 __all__ = [
+    "BoolToVectorSystem",
     "EndEffectorWrenchCalculator",
+    "RigidTransformToVectorSystem",
 ]
