+++ src/brom_drake/utils/leaf_systems/bool_to_vec_system.py
