+++ src/brom_drake/utils/leaf_systems/rigid_transform_to_vector_system.py
